Move Trampoline to common, add test & javadoc.
